Added in image-to-point model training notebook.
From the results of the initial exploration it was clear that data augmentation was not going to work. Added a new notebook to train a image-to-point model based on DeepPicar-v2. Uses a moving average filter to approximate continuous control inputs from the discrete dataset. Needs to loop over the video files to train as cannot all into memory on my laptop.